// src/lib/plugins/types.ts
export interface PluginMenu {
  id: string;
  label: string;
  icon?: string;
  parent?: string;
  position?: string;
  action: string;
  children?: PluginMenu[];
}